+++ src/app/api/links/list/route.ts
@@ -10,6 +10,9 @@ export async function GET(req: Request) {
             return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
         }
 
+        const { searchParams } = new URL(req.url);
+        const { page, limit, search } = Object.fromEntries(searchParams.entries());
+
         let userId: string | null = null;
         try {
             const decoded = jwt.verify(authToken, process.env.JWT_SECRET!) as { id: string };
@@ -20,13 +23,51 @@ export async function GET(req: Request) {
 
         const userPlan = await checkUserPlan(req.headers.get('userPlan') || '');
 
-        const list = await prisma.shortUrl.findMany({
-            where: {
-                userId: userId,
+        let list: any = [];
+        if (search) {
+            list = await prisma.shortUrl.findMany({
+                where: {
+                    userId: userId,
+                    originalUrl: {
+                        contains: search
+                    }
+                }
+            })
+            if (list.length === 0) {
+                list = await prisma.shortUrl.findMany({
+                    where: {
+                        userId: userId,
+                        slug: {
+                            contains: search
+                        }
+                    }
+                })
             }
-        })
+        } else {
+            list = await prisma.shortUrl.findMany({
+                where: {
+                    userId: userId,
+                }
+            })
+        }
 
-        return NextResponse.json({ links: list, allowEdit: userPlan }, { status: 200 });
+        const totalCount = list.length;
+        const pageSize = parseInt(limit || "5", 10);
+        const currentPage = parseInt(page || "1", 10);
+        const totalPages = Math.ceil(totalCount / pageSize);
+        const paginatedList = list.slice((currentPage - 1) * pageSize, currentPage * pageSize);
+
+
+
+        return NextResponse.json({
+            links: {
+                list: paginatedList,
+                totalCount: totalCount,
+                totalPages: totalPages,
+                currentPage: currentPage,
+                pageSize: pageSize
+            }, allowEdit: userPlan
+        }, { status: 200 });
     } catch (error) {
         console.error("Error fetching list:", error);
         return NextResponse.json({ error: "Failed to fetch list" }, { status: 500 });