+++ src/app/edit/page.tsx
@@ -19,7 +19,14 @@ import { Button } from "@/components/ui/button";
 const linkSchema = z.object({
   slug: z.string().min(6, {
     message: "link slug should have at least 6 characters",
-  }),
+  }).max(20, {
+    message: "link slug should have at most 20 characters",
+  }).refine(
+    (value) => /^[a-zA-Z0-9-_]+$/.test(value),
+    {
+      message: "link slug can only contain alphanumeric characters, dashes, and underscores",
+    }
+  ),
   originalUrl: z.string(),
   uses: z.number().min(0).optional(),
   expDate: z.string().optional(),
@@ -189,8 +196,8 @@ function EditPageContent() {
       <Header />
       <main className="flex flex-col gap-[32px] row-start-2 items-center sm:items-start">
         <div className="flex flex-col gap-2 items-center">
-          <h1 className="text-4xl font-bold">edit link</h1>
-          <p className="text-gray-500 text-md mx-2 text-wrap">
+          <h1 className="text-4xl font-bold text-center">edit link</h1>
+          <p className="text-gray-500 text-md mx-2 text-wrap text-center">
             editing link with slug <b>{searchParams.get("slug")}</b>
           </p>
           {isFetching ? (
@@ -207,7 +214,7 @@ function EditPageContent() {
               <Form {...methods}>
                 <form
                   onSubmit={handleSubmit(onSubmit, onError)}
-                  className="flex flex-col gap-2 w-[18rem] py-2"
+                  className="flex flex-col gap-2 w-[20rem] py-2"
                 >
                   <div className="flex flex-col gap-1">
                     <label
@@ -218,7 +225,7 @@ function EditPageContent() {
                     </label>
                     <div className="flex flex-row items-center">
                       <span className="bg-zinc-800 rounded-l-lg border-zinc-500 border-1 px-2 py-2 h-[2.5rem] text-zinc-400">
-                        quickurl.com.br/
+                        quickurl.com.br/r/
                       </span>
                       <Input
                         className="rounded-r-lg rounded-l-none border-zinc-500 h-[2.5rem] py-2 px-2"